% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_moby}
\alias{anno_moby}
\title{anno_moby}
\format{
\subsection{\code{anno_moby}}{

A data frame with 3 elements
\describe{
\item{token}{tibble of tokenized words}
\item{entity}{tibble of entities in text}
\item{document}{tibble of document IDs}
...
}
}
}
\source{
annotated using cleanNLP
}
\usage{
anno_moby
}
\description{
This data frame has annotated text of Moby-Dick from the cleanNLP package
}
\keyword{datasets}
