% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MobyDick}
\alias{MobyDick}
\title{My data}
\format{
\subsection{\code{MobyDick}}{

A data frame with 22243 rows and 1 column:
\describe{
\item{text}{lines of text}
\item{line_number}{line number}
\item{section}{section of work the line is found}
\item{chapter_number}{chapter number}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
MobyDick
}
\description{
This is more descriptive text about my data.
Even more info is provided here ...
}
\keyword{datasets}
