% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{token_type_summary}
\alias{token_type_summary}
\title{token_type_summary}
\format{
\subsection{\code{token_type_summary}}{

A data frame with 135 rows and 18 columns
\describe{
\item{chapter_number}{chapter number}
\item{total}{total number of words}
\item{NOUN_count}{number of nouns}
\item{NOUN_prop}{proportion of nouns}
\item{ADJ_count}{number of adjective}
\item{ADJ_prop}{proportion of adjective}
\item{VERB_count}{number of verbs}
\item{VERB_prop}{proportion of verbs}
\item{ADV_count}{number of adverbs}
\item{ADV_prop}{proportion of adverbs}
\item{PROPN_count}{number of proper nouns}
\item{PROPN_prop}{proportion of proper nouns}
\item{CONJ_count}{number of conjunctions}
\item{CONJ_prop}{proportion of conjunctions}
\item{PRON_count}{number of pronouns}
\item{PRON_prop}{proportion of pronouns}
\item{PUNCT_count}{number of punctuations}
\item{PUNCT_prop}{proportion of punctuations}
...
}
}
}
\source{
data wrangling of annotated text from cleanNLP
}
\usage{
token_type_summary
}
\description{
This package returns a tibble of the number of parts of speech and
their proportion per chapter
}
\keyword{datasets}
