% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{token_with_chapters}
\alias{token_with_chapters}
\title{token_with_chapters}
\format{
\subsection{\code{token_with_chapters}}{

A data frame with 254,474 rows and 11 columns
\describe{
\item{doc_id}{document ID}
\item{sid}{sentence ID}
\item{tid}{token ID}
\item{token}{token}
\item{token_with_ws}{token with whitespace}
\item{lemma}{lemma}
\item{upos}{part of speech}
\item{xpos}{abbreviated part of speech}
\item{tid_source}{start of token set}
\item{relation}{relation to root of token set}
\item{chapter_number}{chapter number}
...
}
}
}
\source{
data wrangling of annotated text from cleanNLP
}
\usage{
token_with_chapters
}
\description{
This data frame returns a tibble of tokenized text with chapter numbers
}
\keyword{datasets}
