% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MobyDick}
\alias{MobyDick}
\title{MobyDick}
\format{
\subsection{\code{MobyDick}}{

A data frame with 22243 rows and 1 column:
\describe{
\item{text}{lines of text}
\item{line_number}{line number}
\item{section}{section of work the line is found}
\item{chapter_number}{chapter number}
...
}
}
}
\source{
\url{https://www.gutenberg.org/ebooks/15}
}
\usage{
MobyDick
}
\description{
This data frame returns a tibble with every line from the text of Moby Dick
along with line number, text section, and chapter number.
}
\keyword{datasets}
